Fixed compiler warnings.
git-svn-id: svn://scribus.net/branches/Version13x/Scribus@7978 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/annota.cpp
@@ -112,7 +112,7 @@ Annota::Annota(QWidget* parent, PageItem *it, int Seite, int b, int h, ScribusVi
 		Pg = new Navigator( GroupBox1, 100, item->annotation().Ziel()+1, view, item->annotation().Extern());
 	else
 	{
-		if (item->annotation().Ziel() < view->Doc->Pages->count())
+		if (static_cast<uint>(item->annotation().Ziel()) < view->Doc->Pages->count())
 			Pg = new Navigator( GroupBox1, 100, item->annotation().Ziel(), view);
 		else
 		{
